rename db service to be less confusing
