+++ src/app/backend/audit-logs/service.ts
@@ -1,6 +1,6 @@
 "use server";
 
-import dbService from "../db/client";
+import dbService from "../db/service";
 
 export type LogEntry = {
   actionType: string;
