+++ src/app/backend/usergroup-management.ts
@@ -5,7 +5,7 @@ import { getSingleUserWithGroupMemberships } from "./user-management";
 import { QCResponse } from "../models/responses/collections";
 import { CustomUserQuery } from "../models/entities/query";
 import { QueryResult } from "pg";
-import dbService from "./db/client";
+import dbService from "./db/service";
 import { adminRequired } from "./db/decorators";
 import { auditable } from "./audit-logs/decorator";
 import { getQueryById } from "./query-building/service";
